find activate voxel functions and scripts
